# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "糖尿病本体 × 患者事实库 融合查询",
    "description": "糖尿病本体与患者事实库融合查询 API。技术验证用途，不是医疗器械，不构成医疗建议。无鉴权、无限流，仅应部署在可信内网或本机。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patients/{pid}/treatment-assessments": {
      "post": {
        "summary": "Treatment Assessment",
        "description": "按 pid 从规范患者表组装镜像，以本地可核验出处生成当前治疗措施的多维度综合报告。",
        "operationId": "treatment_assessment_patients__pid__treatment_assessments_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            },
            "description": "用于组装内部患者镜像的患者标识。",
            "example": "P90002"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Treatment Assessment Patients  Pid  Treatment Assessments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或请求不满足执行条件。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "患者、模板或资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "患者镜像数据库暂不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "x-test-example-title": "按患者镜像生成治疗措施综合评估"
      }
    },
...
    "/patients/{pid}/prediction-demo": {
      "post": {
        "summary": "Prediction Demo",
        "operationId": "prediction_demo_patients__pid__prediction_demo_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            },
            "description": "已播种的合成演示患者标识，仅允许 P91001–P91008。",
            "example": "P91001"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Prediction Demo Patients  Pid  Prediction Demo Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或请求不满足执行条件。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "患者、模板或资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "GraphDB 暂时不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "description": "仅对本模块合成患者执行独立、机器学习模型的数值模型演示，返回输入换算和逐步计算记录。不是机器学习训练或临床疗效预测。",
        "x-test-example-title": "独立运行机器学习模型数值演示"
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Error": {
        "type": "object",
        "description": "标准错误响应。400 也可能表示有意的业务拒绝。",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "description": "错误或业务拒绝原因。部分端点可能返回结构化对象。",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "hint": {
            "type": "string",
            "description": "可选的排障建议。"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://124.223.18.44:8100",
      "description": "远程调试服务"
    }
  ],
  "security": [],
...
}